:zap: reduce data size of motor speed

+++ src/controll/output.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct MotorSpeed {
-    pub left: i16,
-    pub right: i16,
+    pub left: i8,
+    pub right: i8,
 }
